+++ app/Http/Resources/EquipoResource.php
@@ -15,8 +15,9 @@ class EquipoResource extends JsonResource
     public function toArray(Request $request): array
     {
         return [
-            'id'     => $this->id,
-            'nombre' => $this->nombre,
+            'id'      => $this->id,
+            'nombre'  => $this->nombre,
+            'clubId'  => $this->club_id,
         ];
     }
 }